requirements.txt
streamlit
streamlit-chat
python-dotenv
langchain
openai
langchain-core
langchain-openai
langchain-community
huggingface-hub
sentence-transformers
faiss-cpu
pix2text
langsmith